Updated Python module and main script
Main script now takes in variables to use from the input sheets, instead of variables to remove from the input sheets. The EcoPy module is removed from 'analysis', as it is not part of the final product. Furthermore, the Python package PyLab was not needed and is removed from the code, Matplotlib suffices. Several other small changes were made.