Verification added, new structure, updated code.
Moved the sample figures under Source_code. Added the R scripts used for verification. Updated the code of scikit-bio module to allow for CCA and RDA. Added a separate main_script file when using the modules from Scikit_bio_modules for clarity. Added an adapted main_script Python file which was used in the verification in R for the pre-analysis transformations and plotting. In general: all code has been update to their latest versions.